Xml_tree.content_node takes a type parameter
The goal of this change is to enable us to replace the content of a node
with its hash, which will pave the way for cool stuff with irmin. The
point of this patch is not to be merged as-is, but to show the scope of
the changes that are implied by content_node taking a type parameter,
and to serve as a base for discussion.

The annoying side effect of this is that `content` needs to be a
variant. Otherwise we will encountar a cyclic type abbreviation error.

Most changes here are just making sure that the existing functions
receive the arguments they expect (+some formatting by topiary?). There
might be a more elegant solution, I just made the change to the type and
chased down all errors. Worked well enough to get it to compile.

I might send a revision of this patch if I find a way to make it more
elegant.