Replacing addresses with IRIs (major change)
The concept of 'addresses' is being replaced by IRIs
(https://www.ietf.org/rfc/rfc3987.txt). This leads to some big changes
to Forester's behaviour.

Forester's IRI scheme
---------------------

IRIs are a modernisation of URIs that supports unicode characters.
Links in Forester will now point to IRIs; in many cases, these IRIs will
be to external websites, identified with the 'http' or 'https' scheme.
Trees within a forest, however, are addressed by means of an IRI under
the 'forester' scheme. The format for a typical forester IRI is:

    forest://host/addr

Here, 'host' is the global name of the given forest. For instance, I
might set my own host to be 'jonmsterling'; in this case, one might
see an IRI like

    forest://jonmsterling/jms-0001.

The 'host' is configured in forest.toml and then set automatically for
each tree that is compiled as part of the forest.

Although Forester does not yet support mounting external forests, this
change sets the groundwork to being able to refer to trees in external
forests.

Vertices in the Forester graph
------------------------------

The vertices of Forester's graph were previously 'addresses'; vertices
are now *either* IRIs or bits of content. The purpose of supporting bits
of content here is to allow for authorship attributions when the author
in question is not described by a tree, but is instead a blob of content
(as is the case in many bibliographic trees).

Non-automatic resolution of IRIs vs. content
--------------------------------------------

Previously, Forester supported either an address or some content in the
author/contributor attributions. Forester would make a link if the
textual content was an existing address, and it would just emit text
otherwise. This is not tenable in the future, when the address-space
will never be completely known until render-time.

For that reason, \author now requires an IRI; for authors who are not
described by biographical trees in the given forest, you must use
\author/content. The same goes for \contributor.

Handling of tags, taxa, and authors
-----------------------------------

Similarly, tags and taxa are now IRIs, although \tag/content and
\taxon/content are provided similarly to the handling of attributions.
The reason that tags and taxa are IRIs is that the meaning of a given
taxon or tag is often specific to a given forest or even to a specific
part of a forest.

Taxa are rendered using the \title field of the corresponding tree. If
no such tree exists, they are just rendered as the text of their IRI
relative to given forest.

Handling of bibliography queries
--------------------------------

Previously, many queries (including built-in ones, like bibliography)
would filter based on taxon (sc. 'reference'). This is not really
tenable any longer because the use of the 'reference' taxon is a
convention local to a given forest.

To deal with this, unary relations (sets) have been added to Forester's
database and query engine. A reference tree should have the following
declaration in its frontmatter:

   \add-to-set{\set/references}

Then it will be picked up by the built-in bibliography query.