Handle import graph separately from database
There are a few reasons for this, some technical and some
"design"-related.

The import graph is not a good fit for datalog because the main purpose
of the import graph is to be able to run a topological sort. In the
current setup, the import graph is stored as an actual graph data
structure *and* in datalog, and while this is OK, I am thinking that we
don't want to expose it to datalog queries anyway because the latter are
supposed to pertain to the structure of the forest *after* compilation.
With that said, there may be interesting possibilities for query-based
compilation, but I think that this should use a separate datalog
database if we introduce it.

Finally, I think that the import graph should be just part of the forest
state rather than looked up in the table of graphs, because the latter
will disappear as soon as the legacy query engine is removed.