Rewrite the logic of "forester new" (details below)
This resolves two issues:

1. forester/135: the logic to choose the destination of the new tree
   was borked in two ways: first, it was using the location of the last
   tree without chopping off the filename, and second it was apparently
   confusing the order of arguments to List.fold_left. (Very subtle, not
   caught by types!)

2. forester/88: Prefixes ought to be optional.

NOTE FOR KENTO: the behaviour of the test changed in a subtle way that
seems to involve the cram test sandbox. I am not sure what to make of
this, but I promoted the test. Feel free to fix further if something is
broken.

References: https://todo.sr.ht/~jonsterling/forester/135
References: https://todo.sr.ht/~jonsterling/forester/88