change the "Contributions" backmatter query to only consider references
This is an experimental change. Before this change, Contributions would
include every tree the person has authored or contributed to, but that
becomes quickly useless and the number gets higher. What is more useful
to see is a version of this query that returns only trees designated as
"references", e.g. publications, etc.