Defer to HTML in IR for basic things like p, em, etc.
This was a harrowing change due to the very strange and overly
parenthesised formatting of Test_DSL.ml --- and that test doesn't seem
very valuable as-is because it seems to really be testing beta-reduction
and/or my ability to remember what I wrote above; there is no logic
being tested here AFAICT...

This also changes render_context_node in search.Context. Not sure if my
change is correct, but it makes the tests pass.