Switch from iri library to uri
We are agnostic as to the standard; I am switching because I think the
uri library may be better maintained, or at least I have more confidence
that I will be able to get what I need done due to the proximity of
Anil Madhavapeddy to me at work.

Regardless of the library we use for parsing, etc., we need to store our
own *pure* representation in order for things like equality and comparison
to work.