Subdirectories and base URLs in output (breaking change)
The use of relative URLs in the XML output is far too fragile, and
prevents (for example) the emission of subdirectories, etc. The purpose
of being able to emit subdirectories is *not* to allow authors to impose
folder structure, which we find to be counterproductive at the present
time, but rather to allow Forester to place foreign and ephemeral trees
in appropriate subdirectories, etc.

In order to support this, the user has to make a sacrifice:

    Forests can no longer be viewed on the browser via the local file
    system, but must instead be viewed through an HTTP server.

Right now that is easy enough by running

  python3 -m http.server

but we will add a built-in http server to serve the output directory.
Going forward, this will be replaced by the HTMX server as soon as that
is ready.

Configurations are adjusted accordingly: there is a renderer.base_url
field that you can set in case you intend to deploy the rendered forest
at a subdirectory of the host rather than at the root. For example:

   [renderer]
   base_url = "cambridge-disc-maths/"

would be helpful if you want to deploy your forest at
"http://jonmsterling.com/cambridge-disc-maths/".

The forest.home configuration key is moved to renderer.home.