Improve error reporting strategy
We adopt the newly introduced Tree.t type. This reduces the number of
hash tables in the compiler state. A tree now 'remembers' its previous
phases, allowing us to do more sophisticated analysis for the language
server.

The switch from the `Reporter` to the `StructuredReporter` has degraded
the error message quality. I have added an exemplary test in
`lib/core/test` that show how to test the displayed errors. The plan is
to create minimal examples for each error message. At the moment I
suspect that many of these are unreachable. I think this task will
create a lot of clarity regarding the behavior of the compiler when
encountering an error.

Expansion uses topo_iter instead of topo_fold, directly updating the
state and simplifying the expansion procedure. Units are now stored in
the tree.

In the driver, we introduce an explicit `Report_errors` action. This
allows us to adopt different error handling strategies: When
encountering this action during batch compilation, we just display the
diagnostics and quit. When encountering it while running the lsp, we
update the diagnostic store and continue.

Many functions that were part of the `Forest` module are now in `State`.
I think the time is coming where we rename `State` to `Forest`.