Warn when encountering unresolved identifier
Previously we would always fall back to interpreting as a TeX control
sequence. Now we only do this when in math mode. I have added a bunch of
built-in things that automatically resolve to TeX control sequences no
matter the mode, because currently there is no way to control what mode
macro arguments are evaluated in (except making them lazy, which is
probably going to annoy users of forester if we impose it). So this is a
best effort thing.