Undo submodule; vendor the theme latest version (details below)
I realised too late that in order to package for opam via a git tag
archive, we need to vendor it. But this raises the question of how we
should keep this thing up to date in the future; I'm not overly happy
with doing this manually again and again, which is error-prone.